refactor: merge npm scripts
