+++ examples/aspida-node/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "node",
+  "name": "aspida-node",
   "version": "0.0.0",
   "private": true,
   "scripts": {
