+++ examples/browser/package.json
@@ -1,11 +1,10 @@
 {
-  "name": "with-typescript",
+  "name": "browser",
+  "version": "0.0.0",
   "private": true,
   "scripts": {
-    "dev": "npm-run-all mock:build --parallel watch mock:watch",
-    "build": "axios-mock-server --build && tsc",
-    "watch": "nodemon",
-    "type-check": "tsc --noEmit",
+    "dev": "npm-run-all mock:build --parallel dev:server mock:watch",
+    "dev:server": "webpack-dev-server",
     "mock:build": "axios-mock-server --build",
     "mock:watch": "axios-mock-server --watch"
   },
@@ -14,8 +13,9 @@
   },
   "devDependencies": {
     "axios-mock-server": "latest",
-    "nodemon": "^1.19.2",
     "npm-run-all": "^4.1.5",
-    "typescript": "^3.6.3"
+    "webpack": "^4.39.3",
+    "webpack-cli": "^3.3.8",
+    "webpack-dev-server": "^3.8.0"
   }
 }
