+++ examples/browser/public/index.html
