+++ examples/browser/src/index.js
