+++ examples/browser/webpack.config.js
