+++ examples/node/nodemon.json
