+++ examples/node/package.json
@@ -1,0 +1,19 @@
+{
+  "name": "node",
+  "version": "0.0.0",
+  "private": true,
+  "scripts": {
+    "dev": "npm-run-all mock:build --parallel watch mock:watch",
+    "watch": "nodemon",
+    "mock:build": "axios-mock-server --build",
+    "mock:watch": "axios-mock-server --watch"
+  },
+  "dependencies": {
+    "axios": "^0.19.0"
+  },
+  "devDependencies": {
+    "axios-mock-server": "latest",
+    "nodemon": "^1.19.2",
+    "npm-run-all": "^4.1.5"
+  }
+}
