+++ examples/node/src/index.js
