+++ examples/with-nuxtjs/README.md
