+++ examples/with-nuxtjs/jsconfig.json
