+++ examples/with-nuxtjs/nuxt.config.js
