+++ examples/with-nuxtjs/pages/index.vue
