+++ examples/with-nuxtjs/plugins/mock.js
