+++ examples/with-typescript/nodemon.json
