+++ examples/with-typescript/package.json
@@ -1,9 +1,12 @@
 {
-  "name": "node",
+  "name": "with-typescript",
+  "version": "0.0.0",
   "private": true,
   "scripts": {
     "dev": "npm-run-all mock:build --parallel watch mock:watch",
+    "build": "axios-mock-server --build && tsc",
     "watch": "nodemon",
+    "type-check": "tsc --noEmit",
     "mock:build": "axios-mock-server --build",
     "mock:watch": "axios-mock-server --watch"
   },
@@ -13,6 +16,7 @@
   "devDependencies": {
     "axios-mock-server": "latest",
     "nodemon": "^1.19.2",
-    "npm-run-all": "^4.1.5"
+    "npm-run-all": "^4.1.5",
+    "typescript": "^3.6.3"
   }
 }
