+++ examples/with-typescript/src/index.ts
