+++ examples/with-typescript/tsconfig.json
