+++ packages/aspida-mock/__tests__/index.spec.ts
@@ -1,5 +1,5 @@
 import axios, { AxiosInstance } from 'axios'
-import mockServer, { MockRoute, asyncResponse, MockResponse } from '~/src'
+import mockServer, { MockRoute, asyncResponse, MockResponse } from '~/aspida-mock/src'
 
 describe('initialize', () => {
   const mock = mockServer()
