+++ packages/aspida-mock/__tests__/mocks/$mock.js
