+++ packages/aspida-mock/__tests__/mocks/collections/entries.json.ts
@@ -1,4 +1,4 @@
-import { MockMethods } from '~/src/types'
+import { MockMethods } from '~/aspida-mock/src/types'
 
 export const entries = [
   { id: 0, title: 'aaa' },
