+++ packages/aspida-mock/__tests__/mocks/index.ts
@@ -1,4 +1,4 @@
-import { MockMethods } from '~/src/types'
+import { MockMethods } from '~/aspida-mock/src/types'
 
 export default {
   get: () => [200, 'Hello world!']
