+++ packages/aspida-mock/__tests__/mocks/users/self.ts
@@ -1,4 +1,4 @@
-import { MockMethods } from '~/src/types'
+import { MockMethods } from '~/aspida-mock/src/types'
 import { users } from './index'
 
 export default { get: () => [200, users[0]] } as MockMethods
