+++ packages/aspida-mock/docs/ja/README.md
@@ -1,6 +1,6 @@
 <p align="right">
-  <a href="https://github.com/m-mitsuhide/axios-mock-server#readme">🇺🇸English</a> |
-  <a href="https://github.com/m-mitsuhide/axios-mock-server/blob/develop/docs/ja/README.md">🇯🇵日本語</a>
+  <a href="https://github.com/aspidajs/aspida/blob/develop/packages/aspida-mock/README.md">🇺🇸English</a> |
+  <a href="https://github.com/aspidajs/aspida/blob/develop/packages/aspida-mock/docs/ja/README.md">🇯🇵日本語</a>
 </p>
 
 <h1>axios-mock-server</h1>
