--- packages/aspida-mock/examples/browser/package.json
@@ -1,20 +1,0 @@
-{
-  "name": "browser",
-  "private": true,
-  "scripts": {
-    "dev": "npm-run-all mock:build --parallel dev:server mock:watch",
-    "dev:server": "webpack-dev-server",
-    "mock:build": "axios-mock-server --build",
-    "mock:watch": "axios-mock-server --watch"
-  },
-  "dependencies": {
-    "axios": "^0.19.0"
-  },
-  "devDependencies": {
-    "axios-mock-server": "latest",
-    "npm-run-all": "^4.1.5",
-    "webpack": "^4.39.3",
-    "webpack-cli": "^3.3.8",
-    "webpack-dev-server": "^3.8.0"
-  }
-}
