--- packages/aspida-mock/tsconfig.json
@@ -1,19 +1,0 @@
-{
-  "compilerOptions": {
-    "baseUrl": ".",
-    "declaration": true,
-    "declarationMap": true,
-    "downlevelIteration": true,
-    "esModuleInterop": true,
-    "lib": ["dom", "dom.iterable", "es6"],
-    "module": "commonjs",
-    "moduleResolution": "node",
-    "outDir": "dist",
-    "paths": { "~/*": ["./*"] },
-    "resolveJsonModule": true,
-    "sourceMap": true,
-    "strict": true,
-    "target": "es5"
-  },
-  "exclude": ["node_modules", "dist", "examples", "__tests__/lib/mocks/$mock.ts"]
-}
