+++ packages/aspida/__tests__/utils/updateSnapshot.ts
@@ -1,7 +1,7 @@
 import build from '../../src/buildTemplate'
 import write from '../../src/writeRouteFile'
 
-const input = './__tests__/apis'
+const input = 'packages/aspida/__tests__/apis'
 const baseurl = 'https://example.com'
 
 write(build(input, baseurl))
