+++ packages/aspida/tsconfig.build.json
@@ -1,4 +1,5 @@
 {
-  "extends": "./tsconfig.json",
+  "extends": "../../tsconfig.json",
+  "compilerOptions": { "outDir": "dist" },
   "include": ["src/**/*"]
 }
